Make base wand entity class to be derived from, with death control for entity death animations. Make familiars use standard angle limit and interval methods (to be overridden) so body rotation control is only defined in base familiar class. Update some wand textures. Make phylacteries Curios charms. Make move control yaw speed slower.

+++ src/main/java/com/beesechurger/flyingfamiliars/entity/ai/FamiliarBodyRotationControl.java
@@ -14,12 +14,12 @@ public class FamiliarBodyRotationControl extends BodyRotationControl
     private final BaseFamiliarEntity familiar;
 
     private final FFEnumValues.FamiliarMoveTypes rotationType;
-    private final float angleLimit;
-    private final float angleInterval;
+    private final double angleLimit;
+    private final double angleInterval;
     private int headStableTime;
     private float lastStableYHeadRot;
 
-    public FamiliarBodyRotationControl(BaseFamiliarEntity familiar, FFEnumValues.FamiliarMoveTypes rotationType, float angleLimit, float angleInterval)
+    public FamiliarBodyRotationControl(BaseFamiliarEntity familiar, FFEnumValues.FamiliarMoveTypes rotationType, double angleLimit, double angleInterval)
     {
         super(familiar);
         this.familiar = familiar;
